correctly return ssize_t.

+++ lib/libc_r/uthread/uthread_readv.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD$	*/
+/*	$OpenBSD: uthread_readv.c,v 1.3 1999/11/25 07:01:42 d Exp $	*/
 /*
  * Copyright (c) 1995-1998 John Birrell <jb@cimlogic.com.au>
  * All rights reserved.
@@ -45,7 +45,7 @@
 ssize_t
 readv(int fd, const struct iovec * iov, int iovcnt)
 {
-	int	ret;
+	ssize_t	ret;
 	int	type;
 
 	/* Lock the file descriptor for read: */
